+++ crates/gpt2-nvfp4/tests/support/attention_core_scratch.rs
@@ -4,10 +4,10 @@ use rust_kernels_cuda::attention::CausalAttentionBackwardTcScratch;
 
 pub struct AttentionCoreScratchBuffers {
     softmax_d: DeviceBuffer<f32>,
-    q: DeviceBuffer<f32>,
-    k: DeviceBuffer<f32>,
-    v: DeviceBuffer<f32>,
-    d_out: DeviceBuffer<f32>,
+    q: DeviceBuffer<u16>,
+    k: DeviceBuffer<u16>,
+    v: DeviceBuffer<u16>,
+    d_out: DeviceBuffer<u16>,
     scores: DeviceBuffer<f32>,
     dot: DeviceBuffer<f32>,
     p: DeviceBuffer<f32>,
